refactor: move arp mac lookup utilities into infra module
